requirements.txt
aiogram==3.12.0
openai==0.28.0
python-dotenv==0.21.1

